geolocator

+++ lib/screens/note_list_screen.dart
@@ -1,13 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:notes/services/notes_service.dart';
 import 'package:notes/widgets/note_dialog.dart';
+import 'package:url_launcher/url_launcher.dart';
+
 
 class NoteListScreen extends StatefulWidget {
   const NoteListScreen({super.key});
 
+
   @override
   State<NoteListScreen> createState() => _NoteListScreenState();
 }
+
 
 class _NoteListScreenState extends State<NoteListScreen> {
   @override
@@ -33,8 +37,22 @@ class _NoteListScreenState extends State<NoteListScreen> {
   }
 }
 
+
 class NoteList extends StatelessWidget {
   const NoteList({super.key});
+
+
+  Future<void> _launchMaps(double latitude, double longitude) async {
+    Uri googleUrl = Uri.parse(
+        'https://www.google.com/maps/search/?api=1&query=$latitude,$longitude');
+    try {
+      await launchUrl(googleUrl);
+    } catch (e) {
+      print('Could not open the map: $e');
+      // Optionally, show a message to the user
+    }
+  }
+
 
   @override
   Widget build(BuildContext context) {
@@ -57,19 +75,21 @@ class NoteList extends StatelessWidget {
                   child: InkWell(
                     onTap: () {
                       showDialog(
-                          context: context,
-                          builder: (context) {
-                            return NoteDialog(
-                              note: document,
-                            );
-                          });
+                        context: context,
+                        builder: (context) {
+                          return NoteDialog(note: document);
+                        },
+                      );
                     },
                     child: Column(
                       children: [
                         document.imageUrl != null &&
                                 Uri.parse(document.imageUrl!).isAbsolute
                             ? ClipRRect(
-                                borderRadius: BorderRadius.circular(16),
+                                borderRadius: const BorderRadius.only(
+                                  topLeft: Radius.circular(16),
+                                  topRight: Radius.circular(16),
+                                ),
                                 child: Image.network(
                                   document.imageUrl!,
                                   fit: BoxFit.cover,
@@ -82,14 +102,55 @@ class NoteList extends StatelessWidget {
                         ListTile(
                           title: Text(document.title),
                           subtitle: Text(document.description),
-                          trailing: InkWell(
-                            onTap: () {
-                              NoteService.deleteNote(document);
-                            },
-                            child: const Padding(
-                              padding: EdgeInsets.symmetric(vertical: 10),
-                              child: Icon(Icons.delete),
-                            ),
+                          trailing: Row(
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              IconButton(
+                                icon: const Icon(Icons.map),
+                                onPressed: document.latitude != null &&
+                                        document.longitude != null
+                                    ? () {
+                                        _launchMaps(document.latitude!,
+                                            document.longitude!);
+                                      }
+                                    : null, // Disable the button if latitude or longitude is null
+                              ),
+                              InkWell(
+                                onTap: () {
+                                  showDialog(
+                                    context: context,
+                                    builder: (BuildContext context) {
+                                      return AlertDialog(
+                                        title: const Text('Konfirmasi Hapus'),
+                                        content: Text(
+                                            'Yakin ingin menghapus data \'${document.title}\' ?'),
+                                        actions: <Widget>[
+                                          TextButton(
+                                            child: const Text('Cancel'),
+                                            onPressed: () {
+                                              Navigator.of(context).pop();
+                                            },
+                                          ),
+                                          TextButton(
+                                            child: const Text('Hapus'),
+                                            onPressed: () {
+                                              NoteService.deleteNote(document)
+                                                  .whenComplete(() =>
+                                                      Navigator.of(context)
+                                                          .pop());
+                                            },
+                                          ),
+                                        ],
+                                      );
+                                    },
+                                  );
+                                },
+                                child: const Padding(
+                                  padding: EdgeInsets.symmetric(vertical: 10),
+                                  child: Icon(Icons.delete),
+                                ),
+                              ),
+                            ],
                           ),
                         ),
                       ],
